Added fixes with issues in ui
