+++ src/utils/config.ts
@@ -1,8 +1,8 @@
 import { MMKV } from "react-native-mmkv";
 
 // export const BASE_URL = 'http://172.26.171.77:7208';//office lan
-export const BASE_URL = 'http://172.26.171.108:7208'; //office wifi mc1
-// export const BASE_URL = 'http://192.168.1.10:7208';//dorm
+// export const BASE_URL = 'http://172.26.171.108:7208'; //office wifi mc1
+export const BASE_URL = 'http://192.168.55.107:7208';//dorm
 
 
 export const storage = new MMKV();
